move RandomEnumUtils.randomEnum() to RandomUtils().nextEnum()
