+++ src/main/java/org/apache/commons/lang3/RandomUtils.java
@@ -16,6 +16,7 @@
  */
 package org.apache.commons.lang3;
 
+import java.util.List;
 import java.util.Random;
 
 /**
@@ -170,5 +171,17 @@ public class RandomUtils {
         }
         
         return startInclusive + ((endInclusive - startInclusive) * RANDOM.nextFloat());
-    }    
+    }
+
+    /**
+     * Returns a random instance of an Enum type
+     *
+     * @param <E> the type of the enumeration
+     * @param enumClass  the class of the enum to query, not null
+     * @return random instance of an Enum type
+     */
+    public static <E extends Enum<E>> E nextEnum(final Class<E> enumClass) {
+        List<E> enumList = EnumUtils.getEnumList(enumClass);
+        return enumList.get(RandomUtils.nextInt(0, enumList.size()));
+    }
 }
